Created *Window* module; Graphics isn't a class anymore
- Changes to Ball module because of new draw() function; it isn't working yet :/

+++ include/player.hpp
@@ -6,16 +6,16 @@
 class Player
 {
     private:
-        double x, y;                  // Position of the bar's center
+        float x, y;                  // Position of the bar's center
         unsigned int score;           // Number of points obtained
 
     public:
-        constexpr static double v = 2.0;  // Scalar speed
-        constexpr static double rx = 1.0; // Distance from center to left/right
-        constexpr static double ry = 2.0; // Distance from center to up/down
+        constexpr static float v = 2.0;  // Scalar speed
+        constexpr static float rx = 1.0; // Distance from center to left/right
+        constexpr static float ry = 2.0; // Distance from center to up/down
 
         // Creates a bar with center on specified coordinates
-        Player(double x, double y);
+        Player(float x, float y);
 
         // Moves the Player's bar upwards
         void moveUp();
@@ -27,16 +27,16 @@ class Player
         void updateScore();
 
         // Returns the Player's "x" position
-        double getX();
+        float getX();
 
         // Returns the Player's "y" position
-        double getY();
+        float getY();
 
         // Returns the bar's "x" radius
-        double getRx();
+        float getRx();
 
         // Returns the bar's "y" radius
-        double getRy();
+        float getRy();
 
         // Returns the Player's "y" radius
         unsigned int getScore();
